making all the lists but going to pause soon

+++ src/objektOriented/CØØP/Customer.java
@@ -1,23 +1,23 @@
 package objektOriented.CØØP;
 
 public class Customer {
-    private int money;
+    private double money;
     private String kreditkard;
     private String schüperCard;
     private String name;
 
-    public Customer(int money, String kreditkard, String schüperCard, String name) {
+    public Customer(double money, String kreditkard, String schüperCard, String name) {
         this.money = money;
         this.kreditkard = kreditkard;
         this.schüperCard = schüperCard;
         this.name = name;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
