+++ src/objektOriented/CØØP/Groceries/Sweets.java
@@ -2,10 +2,10 @@ package objektOriented.CØØP.Groceries;
 
 public class Sweets {
     private String product;
-    private int price;
+    private double price;
     private boolean barcode;
 
-    public Sweets(String product, int price, boolean barcode) {
+    public Sweets(String product, double price, boolean barcode) {
         this.product = product;
         this.price = price;
         this.barcode = barcode;
@@ -19,11 +19,11 @@ public class Sweets {
         this.product = product;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
